backend/requirements.txt
fastapi
uvicorn[standard]
torch>=2.0
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
pillow
numpy
